Rename PyDAG to PyDigraph and reoganize modules
This commit makes 2 changes, the first is that the rust class PyDAG is
renamed to be PyDiGraph. This makes it more clear that the class can be
used for any directed graph, not just a DAG. The PyDAG class is then
added as a subclass without any changes on the python side of the
module. This will enable a seamless transition for current users where
the only class exposed is PyDAG. Additionally, it will let users
differentiate between a PyDAG and PyDiGraph. Idealling we would want to
only have the cycle checking options exposed for the PyDAG class, but
there isn't any harm if PyDiGraph has it too.

Corresponding with this change the class code for PyDiGraph was moved
to a separate module to make it easier to follow. After this commit
lib.rs only contains the algorithms code and the module configuration.
This makes it easier to follow as lib.rs has been growing as new
algorithms are added.

Fixes #20
