+++ retworkx/__init__.py
@@ -1,1 +1,5 @@
 from .retworkx import *
+
+
+class PyDAG(PyDiGraph):
+    pass
